VarName change, Adapter split
